# docs/mobile-openapi.yaml
openapi: 3.0.3
info:
  title: NGA Admin Mobile API (subset)
  version: "1.0.0"
  description: |
    Staff native clients use Bearer JWT on the same Next.js `/api/*` hosts as the web app.
    Optional header `X-NGA-Client-Version` identifies app builds for server logs.
servers:
  - url: https://{host}
    variables:
      host:
        default: example.com
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate tokens (native body)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken, client]
              properties:
                refreshToken: { type: string }
                client: { type: string, example: native }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
...
  /api/admin/booking-payments:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: booking_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Rows }
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Record payment (admin JWT)
      responses:
        "201": { description: Created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenBundle:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, example: 3600 }
...
    RefreshResponse:
      type: object
      properties:
        success: { type: boolean }
        tokens:
          $ref: "#/components/schemas/TokenBundle"
        message: { type: string }